Added files via upload
I still cannot get this to run properly. I do not know if the problem is that I am using a MAC. Java programs run weird on my mac sometimes and then run fine on windows computers, so maybe this program is fine. Take a look at this and let me know if you can see any problems.